normalize dest name resolution for pageSrc files
There was an issue when slicing abspath based on the length of roodir.
It would only work if pageSrc was defined at a level deeper than the grunt intance.
For example, `pageSrc: '../pages'` would cause broken destination names, e.g. "ex.html" instead of "index.html"
Using node's 'path' module, the path can be normalized for comparison.

Not sure if you want to write unit tests for these cases. Might be worthwhile to use the built-in 'files' as opposed to custom options.

Note that this is probably an issue other places where you are treating filepaths as strings.